Issue : Re-editing the same Point spot on imagebasemap throws an error.
Root cause : Condition fails to enter into imagebasemap section in setSelectedVertexToEdit method, incorrectly assued that, if its a Point, editingModeData.spotToEdit is always empty. This condition fails when the same spot is redited. (Specific to imagebasemap only).

Fix: Added another condition to allow, if editingModeData.spotEdititng is not empty and the spot's type is POINT.