Issue : Re-edit of same Point spot on imagebasemap throws an error.
Root cause: Condition in setSelectedVertexToEdit fails, and doesn't enter the logic for imagebasemap required conversions.
ncorrectly assumed that if the spot is Point (for edit), editingModeData.spotEditing will always be empty, but that fails when same spot is edited twice consecutively.

Fix: added required condition to take care of the scenario