fix(Geography.js): Fixed bug with manual pixel coordinate changes
Pixed coordinates would save as a string if a user manually changed
them in a spot on an imagebasemap. This would lead to a crash because
toFixed would be called when rendering the new coords and the changed
coord would not be an integer and would crash. This part does not go
through the standard form validation.

Closes #433

[skip ci]