fix(useMapsOffline.js): Fixed when ManageOfflineMaps is opened and the mapID gets set to undefined.
This caused an issue when saving project offline mpaas to the device
because the app woul dlook for the mapID, which is the zipID when
saving.   MapId was getting set to undefined because a new map object
was being created when updating the map tiles when opening Manage
Offline Maps.