fix(FormSlider): Fixed bug that kept old value of quality if a Not Recorded value was selected.
When the value of the slider was 6. and there were only 5 choices of the
quality scale, the previous value would save because the quality
key/value would be undefined and therefore be deleted.  Fixed so that
the valus of the quality is set to undefined and then kept as the new
quality.