refactor(Custom Maps): Made a local overlay for saving the map instead of calling the StatusModal using Redux.
Made it so when a user updates the map it only allows them to update the name and overlay areas.
Added an updateCustomMap action so that when a user updates editable options the app is not saving it as a new map and calling the API unnesessarily.
Made the map zoom to the selected custom map and not call the API to save a single map key/value. Calls updateMap instead.

 Also did some minor fomatting and sorting.